Fixed naming discrepancy between package file and package name.
