--- descartes_process_graph/package.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>descartes_path_planning</name>
-  <version>0.0.0</version>
-  <description>The descartes_path_planning package creates joint trajectories for process plans. Process plans are typically underdefined paths through space that allow for processing tolerances, such as unspecified tool roll.</description>
-
-  <maintainer email="dpsolomon@gmail.com">Dan Solomon</maintainer>
-
-  <license>Apache2.0</license>
-  <url type="website">https://github.com/industrial-moveit/descartes_path_planning</url>
-
-  <author email="dpsolomon@gmail.com">Dan Solomon</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-  <build_depend>roscpp</build_depend>
-
-  <run_depend>roscpp</run_depend>
-
-  <export>
-  </export>
-</package>
