+++ descartes_trajectory_planning/include/descartes_trajectory_planning/trajectory_transition.h
