+++ descartes_trajectory_planning/src/trajectory_pt.cpp
@@ -16,14 +16,15 @@
  * limitations under the License.
  */
 /*
- * trajectory_path.cpp
+ * trajectory_pt.cpp
  *
  *  Created on: Jun 5, 2014
  *      Author: Dan Solomon
  */
 
-#include "descartes_path_planning/trajectory.h"
+#include "descartes_trajectory_planning/trajectory_pt.h"
 
 namespace descartes
 {
+
 } /* namespace descartes */
